+++ Authoring/Unity.Tiny.Rendering.Authoring/ExportTinyRenderingSettings.cs
@@ -3,6 +3,7 @@ using Unity.Entities.Runtime.Build;
 using UnityEngine.Assertions;
 using Unity.Tiny.Authoring;
 using Unity.Tiny.Rendering.Settings;
+using UnityEngine;
 
 namespace Unity.Tiny.Authoring
 {
@@ -17,6 +18,14 @@ namespace Unity.Tiny.Authoring
                 Assert.IsTrue(num != 0);
                 var singletonEntity = query.GetSingletonEntity();
                 DisplayInfo di = DisplayInfo.Default;
+
+                // disableSRGB == true means gamma workflow (sRGB disabled).  Otherwise linear (sRGB enabled).
+                // This always comes from the player settings.
+                if (UnityEditor.PlayerSettings.colorSpace == ColorSpace.Gamma)
+                    di.disableSRGB = true;
+                else
+                    di.disableSRGB = false;
+
                 if (buildSettings != null)
                 {
                     if (buildSettings.TryGetComponent<TinyRenderingSettings>(out var settings))
@@ -24,8 +33,7 @@ namespace Unity.Tiny.Authoring
                         di.width = settings.ResolutionX;
                         di.height = settings.ResolutionY;
                         di.autoSizeToFrame = settings.AutoResizeFrame;
-                        di.disableSRGB = settings.DisableSRGB;
-                        di.disableVSync = settings.DisableVsync;   
+                        di.disableVSync = settings.DisableVsync;
                     }
                     else
                         UnityEngine.Debug.LogWarning($"The {nameof(TinyRenderingSettings)} build component is missing from the build setting {buildSettings.name}. Default rendering settings have been exported.");